move components under c.../pages/home
