+++ src/components/pages/home/components/card.tsx
