+++ src/pages/home/index.tsx
@@ -1,6 +1,6 @@
 import React, {FunctionComponent} from 'react'
 import {NextSeo} from 'next-seo'
-import Home from 'components/home'
+import Home from 'components/pages/home'
 
 type HomePageProps = {}
